Update webhook data structure and processing methods
This commit includes modifications to clean up the code and improve data structure for webhook listener. The properties from 'Body' class were moved to 'KiboWebhookData' for better organization and data access. Additionally, in the 'KiboService', the userId was used for the customer_id instead of id. The 'Order' model was also updated to include userId.